chore(db): widen transactions.quantity to support fractional cash-in-lieu quantities
Issue #58 requires the cash-in-lieu SELL transaction produced by a
reverse stock split to carry a fractional quantity (e.g. 0.15 shares),
explicitly as the one deliberate exception to every other BUY/SELL
transaction always having a whole-number quantity. transactions.quantity
was BIGINT/Long, shared by every BUY/SELL row, so it could not represent
that value as written.

This is a genuine contradiction between the issue and the existing
schema, not an ordinary judgment call, so it was raised to the
maintainer before writing the split logic itself; the maintainer chose
widening the shared column over rounding/losing the fractional amount.

- Migrate transactions.quantity from BIGINT to NUMERIC (V9)
- Change Transaction.quantity from Long to BigDecimal
- Fix the two buy()/sell() call sites that set it from a Long request
  field
- Update existing BUY/SELL fixtures and assertions across
  TransactionRepositoryTest, SimulationServiceTest, and
  SimulationControllerIntegrationTest to construct/compare a BigDecimal
  quantity, following this codebase's existing compareTo()-based
  BigDecimal assertion convention. No behavior changes for existing
  whole-number BUY/SELL transactions.

+++ src/backend/src/main/java/com/munehisa/backend/service/SimulationService.java
@@ -180,7 +180,7 @@ public class SimulationService {
         transaction.setAmount(costInBaseCurrency);
         transaction.setTicker(lookup.ticker());
         transaction.setAssetName(lookup.name());
-        transaction.setQuantity(request.quantity());
+        transaction.setQuantity(BigDecimal.valueOf(request.quantity()));
         transactionRepository.save(transaction);
 
         recalculatePositionsAndTotalValue(simulation, positions);
@@ -249,7 +249,7 @@ public class SimulationService {
         transaction.setAmount(proceeds);
         transaction.setTicker(asset.getTicker());
         transaction.setAssetName(asset.getName());
-        transaction.setQuantity(request.quantity());
+        transaction.setQuantity(BigDecimal.valueOf(request.quantity()));
         transactionRepository.save(transaction);
 
         recalculatePositionsAndTotalValue(simulation, positions);
